+++ ExilenceNextApp/src/components/remove-snapshots-dialog/RemoveSnapshotsDialog.styles.ts
@@ -1,0 +1,9 @@
+import makeStyles from '@mui/styles/makeStyles';
+
+const useStyles = makeStyles((theme) => ({
+  dialogActions: {
+    padding: theme.spacing(2),
+  },
+}));
+
+export default useStyles;